LayoutTest platform/mac/webrtc/captureCanvas-webrtc-software-encoder.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=173741
Patch by Youenn Fablet <youenn@apple.com> on 2017-09-21
Reviewed by Alex Christensen.

Making the test more robust by getting the reference canvas to compare with each time a new comparison is done.
Previously, this was done at the start of the comparison.
Also increasing the threshold since we are going from 0 to 255 or vice-versa.

* platform/mac/webrtc/captureCanvas-webrtc-software-encoder.html:

Identifier: 193668@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222361 268f45cc-cd09-0410-ab3c-d52691b4dbfc