Web Inspector: Add autocompletion suggestions for CSS attr based on the selected element's attributes https://bugs.webkit.org/show_bug.cgi?id=177346
Reviewed by Joseph Pecoraro.

Source/WebInspectorUI:

* UserInterface/Controllers/CodeMirrorCompletionController.js:
(WI.CodeMirrorCompletionController.prototype._generateCSSCompletions):
Provide a way for delegates to add completion items for CSS functions based on the name of
the function.

* UserInterface/Views/CSSStyleDeclarationTextEditor.js:
(WI.CSSStyleDeclarationTextEditor.prototype.completionControllerCSSFunctionValuesNeeded):
If the function name is "attr", also return a list of the names of all the selected node's
attributes.

LayoutTests:

* inspector/model/dom-node-expected.txt: Added.
* inspector/model/dom-node.html: Added.


Identifier: 193777@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222483 268f45cc-cd09-0410-ab3c-d52691b4dbfc