Make it possible to easily get verbose ninja output from build-webkit https://bugs.webkit.org/show_bug.cgi?id=177511
Reviewed by Sam Weinig.

* Scripts/build-webkit:
Add a -v/--verbose argument to build-webkit, which sets the environment
variable VERBOSE=1, which buildCMakeGeneratedProject already checks.

* Scripts/webkitdirs.pm:
(buildCMakeGeneratedProject):
Add ninja verbosity arguments to @makeArgs, which are appended
after the -- and thus handed to ninja; the previous implementation
(appending to @args) would only work if @makeArgs was non-empty and thus
the -- was included.


Identifier: 193830@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222538 268f45cc-cd09-0410-ab3c-d52691b4dbfc