[GTK][WPE] Fix playback of GIFs https://bugs.webkit.org/show_bug.cgi?id=176089
Reviewed by Carlos Garcia Campos.

Allow GIFImageReader to decode again already decoded frames. Thanks to this, we don't
need to delete the GIFImageReader when GIFImageDecoder::clearFrameBufferCache() is
called, and the we don't need the lock to avoid crashes in that situation.

Covered by existent tests.

* platform/image-decoders/gif/GIFImageDecoder.cpp:
(WebCore::GIFImageDecoder::clearFrameBufferCache):
(WebCore::GIFImageDecoder::decode):
* platform/image-decoders/gif/GIFImageDecoder.h:
* platform/image-decoders/gif/GIFImageReader.cpp:
(GIFImageReader::decode):


Identifier: 194115@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222836 268f45cc-cd09-0410-ab3c-d52691b4dbfc