Layout Test media/modern-media-controls/scrubber-support/scrubber-support-drag.html is flaky. https://bugs.webkit.org/show_bug.cgi?id=174682
Patch by Antoine Quint <graouts@apple.com> on 2017-10-05
Reviewed by Dean Jackson.

This test could fail since we would start it before the time control was fully sized and only
the scrubber had bounds, but not the elapsed time label. We have to wait on both being fully
sized before starting the test and only then it is reliable.

* media/modern-media-controls/scrubber-support/scrubber-support-drag.html:
* platform/mac/TestExpectations:

Identifier: 194197@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222924 268f45cc-cd09-0410-ab3c-d52691b4dbfc