Make progress toward properly-formed XPC services in the Mac CMake build https://bugs.webkit.org/show_bug.cgi?id=177918
Reviewed by Alex Christensen.

.:

* Source/cmake/WebKitMacros.cmake:
Add a macro wrapper around 'ln'.

Source/WebKit:

* CMakeLists.txt:
* PlatformMac.cmake:
Turn on the storage process on Mac, and fix its build.
Move XPC services to Versions/A/XPCServices and add the toplevel
XPCServices as a symlink of that.
Run the XPC services' Info.plists through configure_file, and substitute
the important variables (most critically the bundle identifier).
Factor out the service setup so we don't keep duplicating it as it
gets more complicated.


Identifier: 194200@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222927 268f45cc-cd09-0410-ab3c-d52691b4dbfc