RenderSVGRoot should check the renderers inside its visualOverflowRect for hit testing if the overflow is visible https://bugs.webkit.org/show_bug.cgi?id=177953 <rdar://problem/34788389>
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-10-05
Reviewed by Simon Fraser.

Source/WebCore:

Test: svg/hittest/svg-visual-overflow-rect.html

If an SVGElement is positioned outside the rectangle of the root element,
it can still be drawn if the root element has "overflow: visible" applied.
But since SVGElement can be drawn in this case, it should be accessible
through nodeAtPoint() which is used for hit testing.

* rendering/svg/RenderSVGRoot.cpp:
(WebCore::RenderSVGRoot::nodeAtPoint):

LayoutTests:

* svg/hittest/svg-visual-overflow-rect-expected.html: Added.
* svg/hittest/svg-visual-overflow-rect.html: Added.

Identifier: 194207@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222934 268f45cc-cd09-0410-ab3c-d52691b4dbfc